pyproject.toml
[tool.poetry]
name = "pact"
version = "0.1.0"
description = ""
authors = ["krystiansiwek <krystian.siwek@yahoo.com>", "sebastiangula <sebastian.gula@gmail.com>", "lukaszptak <lukasz@ulam.io>"]

[tool.poetry.dependencies]
python = "^3.9"
pyteal = "^0.9.0"
py-algorand-sdk = "^1.8.0"
PyYAML = "^5.4.1"
click = "^8.0.3"
python-dotenv = "^0.19.1"
pytealext = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pytest-xdist = "^2.4.0"
hypothesis = "^6.13.10"
black = "^21.11b0"
isort = "^5.10.1"


[tool.black]
exclude = "(.venv|migrations)"
target_version = ["py39"]
line-length = 100

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip = [".venv", "migrations"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
